ZOH-172 - Add event status switcher

+++ src/scandipwa/app/plugin/event-dispatches/cartItemChangeQuantity.plugin.js
@@ -2,13 +2,13 @@ import Event, {
     EVENT_GTM_PRODUCT_ADD_TO_CART,
     EVENT_GTM_PRODUCT_REMOVE_FROM_CART
 } from '../../util/Event';
+import { EVENT_ADD_TO_CART, EVENT_REMOVE_FROM_CART } from '../../component/GoogleTagManager/GoogleTagManager.component';
+import { isEventEnabled } from '../../util/EventConfig';
 
 class QuantityChangePlugin {
-    /*
-        handleChangeState = {};
-    */
-
     handleChangeQuantity = (args, callback, instance) => {
+        if (!isEventEnabled(EVENT_ADD_TO_CART)) return callback(...args);
+
         const [quantity] = args;
         const { item, item: { qty } } = instance.props;
 
@@ -19,32 +19,38 @@ class QuantityChangePlugin {
         };
 
         callback(...args);
-    }
+    };
 
     changeItemQty = (args, callback, instance) => {
+        if (!isEventEnabled(EVENT_ADD_TO_CART)) return callback(...args);
+
         const { newQuantity, item, oldQuantity } = this.handleChangeState;
 
         return callback(...args)
             .then(
                 result => {
                     if (oldQuantity < newQuantity) {
-                        Event.dispatch(EVENT_GTM_PRODUCT_ADD_TO_CART, {
-                            product: item,
-                            quantity: newQuantity - oldQuantity,
-                            isItem: true,
-                            isFromCart: true
-                        });
+                        if (isEventEnabled(EVENT_ADD_TO_CART)) {
+                            Event.dispatch(EVENT_GTM_PRODUCT_ADD_TO_CART, {
+                                product: item,
+                                quantity: newQuantity - oldQuantity,
+                                isItem: true,
+                                isFromCart: true
+                            });
+                        }
                     } else {
-                        Event.dispatch(EVENT_GTM_PRODUCT_REMOVE_FROM_CART, {
-                            item,
-                            quantity: oldQuantity - newQuantity
-                        });
+                        if (isEventEnabled(EVENT_REMOVE_FROM_CART)) {
+                            Event.dispatch(EVENT_GTM_PRODUCT_REMOVE_FROM_CART, {
+                                item,
+                                quantity: oldQuantity - newQuantity
+                            });
+                        }
                     }
 
-                    return result
+                    return result;
                 }
             );
-    }
+    };
 }
 
 const {
@@ -63,4 +69,4 @@ export default {
             'changeItemQty': changeItemQty
         }
     }
-}
+};
